+++ src/openassetio-python/cmodule/src/hostApi/ManagerBinding.cpp
@@ -154,6 +154,31 @@ void registerManager(const py::module& mod) {
           py::arg("entityReference"), py::arg("relationshipTraitsDatas"),
           py::arg("context").none(false), py::arg("successCallback"), py::arg("errorCallback"),
           py::arg("resultTraitSet") = trait::TraitSet{})
+      .def("getWithRelationshipPaged",
+           py::overload_cast<const EntityReferences&, const TraitsDataPtr&, std::size_t,
+                             const ContextConstPtr&,
+                             const Manager::PagedRelationshipSuccessCallback&,
+                             const Manager::BatchElementErrorCallback&, const trait::TraitSet&>(
+               &Manager::getWithRelationshipPaged),
+           py::arg("entityReferences"), py::arg("relationshipTraitsData").none(false),
+           py::arg("pageSize"), py::arg("context").none(false), py::arg("successCallback"),
+           py::arg("errorCallback"), py::arg("resultTraitSet") = trait::TraitSet{})
+      .def(
+          "getWithRelationshipsPaged",
+          [](Manager& self, const EntityReference& entityReference,
+             const trait::TraitsDatas& relationshipTraitsDatas, size_t pageSize,
+             const ContextConstPtr& context,
+             const Manager::PagedRelationshipSuccessCallback& successCallback,
+             const Manager::BatchElementErrorCallback& errorCallback,
+             const trait::TraitSet& resultTraitSet) {
+            validateTraitsDatas(relationshipTraitsDatas);
+            self.getWithRelationshipsPaged(entityReference, relationshipTraitsDatas, pageSize,
+                                           context, successCallback, errorCallback,
+                                           resultTraitSet);
+          },
+          py::arg("entityReference"), py::arg("relationshipTraitsDatas"), py::arg("pageSize"),
+          py::arg("context").none(false), py::arg("successCallback"), py::arg("errorCallback"),
+          py::arg("resultTraitSet") = trait::TraitSet{})
       .def("preflight",
            static_cast<void (Manager::*)(
                const EntityReferences&, const trait::TraitSet&, const ContextConstPtr&,